- GPU Driven: Use InterlockedMax to prevent z-buffer race condition * fix uv mapping error in CPU texture sample + add a default grid texture for default material - add tolerant for zero check

+++ URasterizer/Assets/URasterizer/Codes/GPURasterizer/Shaders/URCommon.hlsl
@@ -25,5 +25,5 @@ StructuredBuffer<uint3> triangleBuffer; // All triangles of the mesh
 
 RWStructuredBuffer<VertexOutBuf> vertexOutBuffer;
 RWTexture2D<float4> frameColorTexture;
-RWTexture2D<float> frameDepthTexture;
+RWTexture2D<uint> frameDepthTexture;
